# rope-manager-backend/openapi.yaml
openapi: 3.0.3
info:
  title: 绳包管理器后端 API
  description: |
    绳包管理器后端服务的完整API文档
    
    ## 认证方式
    系统支持两种认证方式，按优先级顺序：
    1. **Authorization 头部认证（推荐）**: `Authorization: Bearer <token>`
    2. **Cookie 认证**: `Cookie: auth_token=<token>`
    
    ## 权限等级
    - **普通用户 (user)**: 基础功能使用权限
    - **元老 (elder)**: 拥有部分管理权限，可审核资源
    - **版主 (moderator)**: 拥有内容管理权限
    - **管理员 (admin)**: 拥有所有系统权限
    
    ## API 响应格式
    所有API统一返回格式：
    ```json
    {
      "code": 0,
      "message": "success",
      "data": {}
    }
    ```
    - code: 状态码，0表示成功，其他表示错误
    - message: 响应消息
    - data: 响应数据（可选）
    
  version: 1.0.0
  contact:
    name: 绳包管理器API支持
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:15201/api/v1
    description: 开发服务器 - API端点
  - url: http://127.0.0.1:15201
    description: 开发服务器 - 健康检查端点
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，0表示成功
        message:
          type: string
          description: 响应消息
        data:
          description: 响应数据（可选）
      required:
        - code
        - message

    # 用户相关模型
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        target_type:
          type: string
          enum: [Package, User, System]
        target_id:
          type: integer
        content:
          type: string
        status:
          type: string
          enum: [Active, Hidden, Deleted]
        parent_id:
          type: integer
          nullable: true
        likes:
          type: integer
        dislikes:
          type: integer
        pinned:
          type: boolean
        author_name:
          type: string
          nullable: true
        username:
          type: string
          nullable: true
        author_role:
          type: string
          nullable: true
        author_avatar:
          type: string
          nullable: true
        author_qq:
          type: string
          nullable: true
        target_title:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCommentRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        target_type:
          type: string
          enum: [Package, User, System]
        target_id:
          type: integer
        parent_id:
          type: integer
          nullable: true
      required:
        - content
        - target_type
        - target_id
    UpdateCommentRequest:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
        status:
          type: string
          enum: [Active, Hidden, Deleted]
    CommentListResponse:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        total:
          type: integer
          format: int64
        page:
          type: integer
        size:
          type: integer

    # 帖子相关模型
...
    BatchIdsRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: integer
          minItems: 1
      required:
        - ids
    LikeRequest:
      type: object
      properties:
        like:
          type: boolean
      required:
        - like
    DislikeRequest:
      type: object
      properties:
        dislike:
          type: boolean
      required:
        - dislike
    PinRequest:
      type: object
      properties:
        pinned:
          type: boolean
      required:
        - pinned

    # 系统统计相关模型
...
paths:
...
  /users/{id}/comments:
    get:
      tags:
        - 用户管理
      summary: 获取用户评论
      description: 获取指定用户的所有评论
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'

  # 绳包管理接口
...
  /packages/{id}/comments:
    get:
      tags:
        - 绳包管理
      summary: 获取绳包评论
      description: 获取指定绳包的评论列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'
...
  /comments:
    get:
      tags:
        - 评论系统
      summary: 获取评论列表
      description: 获取评论列表（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [Active, Hidden, Deleted]
        - name: target_type
          in: query
          schema:
            type: string
            enum: [Package, User, System]
        - name: target_id
          in: query
          schema:
            type: integer
        - name: user_id
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'
        '403':
          description: 权限不足

    post:
      tags:
        - 评论系统
      summary: 创建评论
      description: 创建新评论
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comment_id:
                            type: integer
        '401':
          description: 未授权
...
  /comments/batch-delete:
    delete:
      tags:
        - 评论系统
      summary: 批量删除评论
      description: 批量删除评论（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchIdsRequest'
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
  /comments/{id}:
    get:
      tags:
        - 评论系统
      summary: 获取评论详情
      description: 根据ID获取评论详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Comment'
        '404':
          description: 评论不存在

    put:
      tags:
        - 评论系统
      summary: 更新评论
      description: 更新评论内容（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: 更新成功
        '403':
          description: 权限不足
        '404':
          description: 评论不存在

    delete:
      tags:
        - 评论系统
      summary: 删除评论
      description: 删除评论（作者或管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
        '403':
          description: 权限不足
        '404':
          description: 评论不存在
  /comments/{id}/replies:
    get:
      tags:
        - 评论系统
      summary: 获取评论回复
      description: 获取指定评论的回复列表
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CommentListResponse'

    post:
      tags:
        - 评论系统
      summary: 回复评论
      description: 回复指定评论
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 2000
              required:
                - content
      responses:
        '200':
          description: 回复成功
        '401':
          description: 未授权
        '404':
          description: 原评论不存在
  /comments/{id}/like:
    post:
      tags:
        - 评论系统
      summary: 点赞/取消点赞评论
      description: 点赞或取消点赞评论
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LikeRequest'
      responses:
        '200':
          description: 操作成功
        '401':
          description: 未授权
        '404':
          description: 评论不存在
  /comments/{id}/dislike:
    post:
      tags:
        - 评论系统
      summary: 点踩/取消点踩评论
      description: 点踩或取消点踩评论
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DislikeRequest'
      responses:
        '200':
          description: 操作成功
        '401':
          description: 未授权
        '404':
          description: 评论不存在
  /comments/{id}/pin:
    post:
      tags:
        - 评论系统
      summary: 置顶/取消置顶评论
      description: 置顶或取消置顶评论（管理员权限）
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinRequest'
      responses:
        '200':
          description: 操作成功
        '403':
          description: 权限不足
        '404':
          description: 评论不存在

  # 帖子管理接口
...
  /community/comments/{package_id}:
    get:
      tags:
        - 社区动态
      summary: 获取资源评论
      description: 获取指定资源的评论列表
      parameters:
        - name: package_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: '#/components/schemas/Comment'
                          total:
                            type: integer
                          page:
                            type: integer
                          size:
                            type: integer
        '500':
          description: 服务器错误

    post:
      tags:
        - 社区动态
      summary: 创建评论
      description: 为指定资源创建评论
      parameters:
        - name: package_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '200':
          description: 评论成功
        '400':
          description: 评论失败
...